+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+# WIP
+
+- Trigger additional `app.user.logout.after` on user logout
+
+
 ## 2.11.4 (2025-07-01)
 
 - Assets: add video transcoding helper function